[chore] 파일 정리
