+++ SMASHING/Presentation/Onboarding/GenderViewController.swift
@@ -30,133 +30,3 @@ final class GenderViewController: BaseViewController {
         genderView.configure(action: action)
     }
 }
-
-final class GenderView: BaseUIView {
-    
-    // MARK: - Properties
-    
-    private var action: ((Gender) -> Void)?
-
-    // MARK: - UI Components
-    
-    private let stackView = UIStackView().then {
-        $0.axis = .horizontal
-        $0.distribution = .fillEqually
-        $0.spacing = 11
-    }
-    
-    private let maleButton = GenderButton(gender: .male)
-    private let femaleButton = GenderButton(gender: .female)
-    
-    // MARK: - Setup Methods
-    
-    override func setUI() {
-        addSubview(stackView)
-        [maleButton, femaleButton].forEach { stackView.addArrangedSubview($0) }
-        
-        maleButton.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(maleTapped)))
-        femaleButton.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(femaleTapped)))
-    }
-    
-    override func setLayout() {
-        stackView.snp.makeConstraints {
-            $0.top.leading.trailing.equalToSuperview()
-            $0.height.equalTo(stackView.snp.width).multipliedBy(0.5).offset(-11)
-        }
-    }
-    
-    func configure(action: @escaping (Gender) -> Void) {
-        self.action = action
-    }
-    
-    // MARK: - Actions
-    
-    @objc private func maleTapped() {
-        handleSelection(gender: .male)
-    }
-    
-    @objc private func femaleTapped() {
-        handleSelection(gender: .female)
-    }
-    
-    private func handleSelection(gender: Gender) {
-        maleButton.isSelected = (gender == .male)
-        femaleButton.isSelected = (gender == .female)
-        action?(gender)
-    }
-}
-
-final class GenderButton: BaseUIView {
-    
-    // MARK: - Properties
-    
-    private let gender: Gender
-    
-    var isSelected: Bool = false {
-        didSet { updateStyle() }
-    }
-
-    // MARK: - UI Components
-    
-    private let contentStackView = UIStackView().then {
-        $0.axis = .vertical
-        $0.alignment = .center
-        $0.spacing = 8
-        $0.isUserInteractionEnabled = false
-    }
-    
-    private let imageView = UIImageView().then {
-        $0.contentMode = .scaleAspectFit
-    }
-    
-    private let label = UILabel().then {
-        $0.font = .pretendard(.textMdM)
-    }
-
-    // MARK: - Init
-    
-    init(gender: Gender) {
-        self.gender = gender
-        super.init(frame: .zero)
-        setupAttributes()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    private func setupAttributes() {
-        self.layer.cornerRadius = 10
-        self.layer.borderWidth = 1
-        self.label.text = gender.name
-        updateStyle()
-    }
-
-    override func setUI() {
-        addSubview(contentStackView)
-        [imageView, label].forEach { contentStackView.addArrangedSubview($0) }
-    }
-
-    override func setLayout() {
-        contentStackView.snp.makeConstraints {
-            $0.center.equalToSuperview()
-        }
-        imageView.snp.makeConstraints {
-            $0.size.equalTo(24)
-        }
-    }
-
-    private func updateStyle() {
-        if isSelected {
-            self.backgroundColor = .Background.selected
-            self.layer.borderColor = UIColor.clear.cgColor
-            self.imageView.image = gender.imageLg.tinted(with: .Text.primaryReverse)
-            self.label.textColor = .Text.primaryReverse
-        } else {
-            self.backgroundColor = .clear
-            self.layer.borderColor = UIColor.Border.secondary.cgColor
-            self.imageView.image = gender.imageLg
-            self.label.textColor = .Text.primary
-        }
-    }
-}
